# api-gateway/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Learning Platform API
  description: API для управления образовательной платформой
  version: '1.0'
servers:
  - url: http://localhost:8080
    description: Generated server url
...
paths:
  /user/mentor/register:
    post:
      tags:
        - User Info Controller
      summary: Выдает пользователю роль МЕНТОР
      description: Позволяет выдать пользователю роль МЕНТОР
      operationId: assignMentorRole
      responses:
        '200':
          description: Роль МЕНТОР выдана
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                type: object
  /slot/create:
    post:
      tags:
        - Time Slot Controller
      summary: Создать слот
      description: Позволяет создать новый слот. Требуются права ADMIN или MENTOR
      operationId: createSlot
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MentorTimeSlotCreateRequest'
        required: true
      responses:
        '200':
          description: Слот успешно создан
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
  /slot/book:
    post:
      tags:
        - Time Slot Controller
      summary: Забронировать слот
      description: Позволяет забронировать слот. Требуются права USER
      operationId: bookSlot
      parameters:
        - name: timeSlotId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Слот успешно забронирован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
...
  /user/{userId}:
    get:
      tags:
        - User Info Controller
      summary: Получить информацию о другом пользователе
      description: Позволяет получить пользовательскую информацию о другом пользователе
      operationId: getOtherUserInfo
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Пользовательская информация выдана
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                type: object
  /user/me:
    get:
      tags:
        - User Info Controller
      summary: Получить свою информацию
      description: Позволяет получить пользовательскую информацию от пользователя
      operationId: getMyUserInfo
      responses:
        '200':
          description: Пользовательская информация выдана
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                type: object
  /slot:
    get:
      tags:
        - Time Slot Controller
      summary: Выдать информацию обо всех слотах ментора
      description: Выдать информацию о слотах. Требуются права USER
      operationId: getMentorSlotsInfoForUser
      parameters:
        - name: mentorId
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Выдан список слотов
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorTimeSlotInfoForUserDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorTimeSlotInfoForUserDto'
  /slot/my:
    get:
      tags:
        - Time Slot Controller
      summary: Выдать информацию обо всех слотах ментора
      description: Выдать информацию о слотах и об участниках. Требуются права MENTOR
      operationId: getMentorSlotsInfo
      responses:
        '200':
          description: Выдан список слотов
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MentorTimeSlotDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorSlotInfoDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MentorSlotInfoDto'
  /progress/course/{courseId}/users:
    get:
      tags:
        - Progress Controller
      summary: Получить всех учеников в курсе
      description: Позволяет получить информацию о всех учениках в курсе с текущим модулем. Требуются права ADMIN или MENTOR
      operationId: getAllUsersAtCourse
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информацию о всех учениках в курсе
          content:
            '*/*':
              schema:
                allOf:
                  - $ref: '#/components/schemas/MenteeProgressDto'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenteeProgressDto'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenteeProgressDto'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MenteeProgressDto'
  /progress/course/{courseId}/statistics:
    get:
      tags:
        - Progress Controller
      summary: Получить прогресс по курсу
      description: Позволяет получить прогресс обучения в курсе автором этого курса. Требуются права ADMIN или MENTOR
      operationId: getCourseProgressByMentor
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Информации о прогрессе учеников
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseProgressResponse'
        '400':
          description: Невалидные входные данные
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseProgressResponse'
        '401':
          description: Не авторизован
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseProgressResponse'
        '403':
          description: Доступ запрещен
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CourseProgressResponse'
...
  /admin/user/{userId}:
    get:
      tags:
        - Admin User Info Controller
      summary: Получить информацию о другом профиле.
      description: Позволяет получить информацию о профиле другого пользователя.
      operationId: adminGetOtherUserInfo
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Пользовательская информация выдана
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                type: object
  /admin/user/me:
    get:
      tags:
        - Admin User Info Controller
      summary: Получить информацию о своем профиле.
      description: Позволяет получить информацию о своем профиле.
      operationId: adminGetMyUserInfo
      responses:
        '200':
          description: Пользовательская информация выдана
          content:
            '*/*':
              schema:
                type: object
        '404':
          description: Пользователь не найден
          content:
            '*/*':
              schema:
                type: object
...
components:
  schemas:
    MentorTimeSlotDto:
      type: object
      properties:
        id:
          type: integer
          description: ID слота
          format: int64
          example: 1
        mentorId:
          type: integer
          description: ID ментора
          format: int64
          example: 1
        rqUId:
          type: string
          description: UUID запроса
          example: 6e8f4e02-c91c-465f-b22d-7f102fca381b
        startTime:
          type: string
          description: Начало открытого слота
          format: date-time
        endTime:
          type: string
          description: Окончание открытого слота
          format: date-time
        slotType:
          type: string
          description: Тип слота
          example: INDIVIDUAL
          enum:
            - INDIVIDUAL
            - GROUP
        slotMeetingType:
          type: string
          description: Тип планируемой встречи
          example: COMMUNICATION
          enum:
            - ACQUAINTANCE
            - COMMUNICATION
            - ACCEPTING
        maxParticipants:
          type: integer
          description: Максимальное число участников
          format: int32
          example: 10
        meetingLink:
          type: string
          description: Ссылка на встречу
          example: https://www.meet.ru/<uuid>
        description:
          type: string
          description: Описание слота
          example: Знакомство и обсуждение плана дальнейшего взаимодействия
        createdAt:
          type: string
          description: Дата и время создания создания слота
          format: date-time
        isActive:
          type: boolean
          description: Признак активности слота
          example: true
      description: Информация о слоте
    MentorTimeSlotCreateRequest:
      required:
        - meetingLink
        - slotMeetingType
        - slotType
      type: object
      properties:
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        slotType:
          type: string
          enum:
            - INDIVIDUAL
            - GROUP
        slotMeetingType:
          type: string
          enum:
            - ACQUAINTANCE
            - COMMUNICATION
            - ACCEPTING
        maxParticipants:
          maximum: 50
          minimum: 1
          type: integer
          format: int32
        meetingLink:
          type: string
        description:
          type: string
...
    UserInfoDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        role:
          type: string
          enum:
            - USER
            - ADMIN
            - MENTOR
        firstName:
          type: string
        lastName:
          type: string
        tgNickname:
          type: string
        tgChatId:
          type: integer
          format: int64
      description: ID автора курса
      example: 5
...
    MentorTimeSlotInfoForUserDto:
      type: object
      properties:
        mentorTimeSlotDto:
          $ref: '#/components/schemas/MentorTimeSlotDto'
        slotFull:
          type: boolean
      description: Информация о слоте для ученика
    MentorSlotInfoDto:
      type: object
      properties:
        slotDto:
          $ref: '#/components/schemas/MentorTimeSlotDto'
        participants:
          type: array
          description: Информация об участниках слота
          items:
            $ref: '#/components/schemas/UserInfoDto'
      description: Подробная информация о слоте ментора и об участниках в нем
    MenteeProgressDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        firstName:
          type: string
        lastName:
          type: string
        currentModuleId:
          type: integer
          format: int64
        tgNickname:
          type: string
    CourseProgressResponse:
      type: object
      properties:
        courseId:
          type: integer
          format: int64
        courseTitle:
          type: string
        mentee:
          type: array
          items:
            $ref: '#/components/schemas/MenteeProgressDto'
        statistic:
          $ref: '#/components/schemas/CourseProgressStatisticDto'
    CourseProgressStatisticDto:
      type: object
      properties:
        totalMenteeCount:
          type: integer
          format: int32
        moduleDistribution:
          type: object
          additionalProperties:
            type: integer
            format: int32
...